+++ include/mylang/CodeGen.h
@@ -266,6 +266,10 @@ private:
     llvm::Function* runtime_chr_ = nullptr;
     llvm::Function* runtime_ord_ = nullptr;
 
+    // ---- RTTI (§五-4) ----
+    llvm::Function* runtime_type_id_ = nullptr;
+    llvm::Function* runtime_type_name_ = nullptr;
+
     // ---- String to double ----
     llvm::Function* runtime_atof_ = nullptr;